backend/requirements.txt
django==3.1.1
djangorestframework==3.11.1
django-rest-knox==4.1.0
django-cors-headers==3.5.0
requests==2.24.0
mysqlclient==2.0.1